pyproject.toml
[project]
name = "phototakeoutsorter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pillow>=11.3.0",
    "pillow-heif>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
